# openapi/road-score.yaml
openapi: 3.0.3
info:
  title: Road score API
  description: |
    Score a road/street image from **1** (best — good street to use) to **100** (worst).
    Responses include a short **rationale**, a longer **explanation**, and structured **analysis**
    (pavement, visibility, hazards, context). Backend uses **OpenRouter** vision models.
  version: 1.4.0

servers:
  - url: http://localhost:8000
    description: Local (uvicorn)
  - url: https://api.example.com
    description: Production (replace with your host)

tags:
  - name: road-score
    description: Road quality from images
  - name: media-seed
    description: Latest crawl export from media seed crawler

paths:
  /health:
    get:
      summary: Liveness check
      operationId: healthCheck
      tags: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /v1/road-score:
    post:
      summary: Score road quality from an image
      description: |
        Upload a single image. Returns **score** (1–100), **rationale** (one line),
        **explanation** (multi-sentence), and **analysis** (structured observations).
      operationId: postRoadScore
      tags:
        - road-score
      parameters:
        - name: model
          in: query
          required: false
          description: Override OpenRouter model id (optional; server has a default).
          schema:
            type: string
            example: openai/gpt-4o-mini
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Road or street photo (JPEG, PNG, or WebP).
      responses:
        "200":
          description: Score, explanation, and analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadScoreResponse"
        "400":
          description: Invalid file type or oversized image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server misconfiguration (e.g. missing API key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: OpenRouter error or unparseable model output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/media-seed/latest:
    get:
      summary: Latest route scores from media seed export
      description: |
        Returns **routes** (ordered list) and **routes_map** (name → score + reason). Each **score** is 1–100
        (1 = best route to use, 100 = worst). **reason** comes from OpenRouter explanation/rationale when the
        export has scored `media_downloads`. Use `full=true` to include the raw export object in **export**.
        Default file: `media_seed_export.json` at repo root; override with `MEDIA_SEED_EXPORT_PATH`.
      operationId: getMediaSeedLatest
      tags:
        - media-seed
      parameters:
        - name: full
          in: query
          required: false
          description: If true, include the full parsed export JSON in `export`.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Route scores with reasons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaSeedLatestResponse"
        "404":
          description: Export file missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Read error or invalid JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    RoadAnalysis:
      type: object
      description: Structured observations visible in the image.
      properties:
        pavement_surface:
          type: string
          description: Surface, markings, cracks, potholes, lane clarity.
          example: "Asphalt in fair condition; visible lane markings; no major potholes in frame."
        visibility_environment:
          type: string
          description: Sight lines, lighting, weather, visual clutter.
          example: "Daylight; generally open view along the road axis."
        hazards_constraints:
          type: string
          description: Obstructions, water, construction, geometry, debris.
          example: "No obvious flooding or large obstacles in the photographed section."
        scene_context:
          type: string
          description: Urban/rural cues, sidewalks, crossings, setting.
          example: "Urban two-lane street with buildings along one side."

    RoadScoreResponse:
      type: object
      required:
        - score
        - analysis
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 100
          description: 1 = best street to use, 100 = worst.
          example: 24
        rationale:
          type: string
          nullable: true
          description: One-line summary for the score.
          example: "Good pavement and markings with no major hazards visible."
        explanation:
          type: string
          nullable: true
          description: Multi-sentence narrative for end users.
          example: "The road surface appears even with clear lane paint. Visibility along the corridor looks adequate for the conditions shown. No severe obstacles are visible in this frame, though the assessment is limited to what the photo shows."
        analysis:
          $ref: "#/components/schemas/RoadAnalysis"
        model:
          type: string
          nullable: true
          description: Vision model id used on OpenRouter.
          example: openai/gpt-4o-mini

    RouteScoreValue:
      type: object
      required:
        - score
        - reason
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 100
          description: 1 = best route to use, 100 = worst.
        reason:
          type: string
          description: Why this score (model text when present in export).

    RouteScoreEntry:
      type: object
      required:
        - name
        - score
        - reason
      properties:
        name:
          type: string
          description: Route label / export key (corridor text, image:file, audio:file, …).
        score:
          type: integer
          minimum: 1
          maximum: 100
          description: 1 = best route to use, 100 = worst.
        reason:
          type: string
          description: Why this score (model text when present in export).

    MediaSeedLatestResponse:
      type: object
      required:
        - routes
        - routes_map
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteScoreEntry"
          description: Ordered entries (from score_keys_this_run when present).
        routes_map:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RouteScoreValue"
          description: Same data keyed by route name.
        crawled_at:
          type: string
          nullable: true
          description: From export `_meta.crawled_at`.
        source_url:
          type: string
          nullable: true
          description: From export `_meta.url`.
        export:
          type: object
          nullable: true
          additionalProperties: true
          description: Full export root when `full=true`.

    Error:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items: {}